refactor: MultiSplitView 상태 모델 클래스 분리
